Made several separate classes for benchmarks results and json report.
