+++ performance/src/main/kotlin-native/org/jetbrains/ring/Utils.kt
@@ -51,9 +51,9 @@ class Random() {
 
 //-----------------------------------------------------------------------------//
 
-class FileWriter() {
-    companion object {
-        fun writeToFile(fileName: String, text: String): Unit {
+actual class FileWriter() {
+    actual companion object {
+        actual fun writeToFile(fileName: String, text: String): Unit {
             val file = fopen(fileName, "wt")
             if (file == null) throw Error("Cannot write file '$fileName'")
             try {
