+++ performance/src/main/kotlin/main.kt
@@ -15,6 +15,7 @@
  */
 
 import org.jetbrains.ring.Launcher
+import org.jetbrains.ring.JsonReportCreator
 
 fun main(args: Array<String>) {
     var numWarmIterations = 0       // Should be 100000 for jdk based run
@@ -34,12 +35,14 @@ fun main(args: Array<String>) {
             jsonReport = args[2].toString()
         }
         else -> {
-            println("Usage: perf [# warmup iterations] [# attempts]")
+            println("Usage: perf [# warmup iterations] [# attempts] [# path of json report]")
             return
         }
     }
 
     println("Ring starting")
     println("  warmup  iterations count: $numWarmIterations")
-    Launcher(numWarmIterations, numberOfAttempts, jsonReport).runBenchmarks()
+    val results = Launcher(numWarmIterations, numberOfAttempts).runBenchmarks()
+    if (jsonReport != null)
+        JsonReportCreator(results).printJsonReport(jsonReport)
 }
